refactor: datatypes
