+++ kolena/workflow/_validators.py
@@ -21,12 +21,12 @@ from typing import Type
 from typing import Union
 
 from kolena._experimental.workflow.thresholded import ThresholdedMetrics
+from kolena._utils.datatypes import DATA_TYPE_FIELD
+from kolena._utils.datatypes import DataObject
 from kolena._utils.datatypes import get_args
 from kolena._utils.datatypes import get_origin
 from kolena.errors import WorkflowMismatchError
 from kolena.workflow._datatypes import _SCALAR_TYPES
-from kolena.workflow._datatypes import DATA_TYPE_FIELD
-from kolena.workflow._datatypes import DataObject
 from kolena.workflow.annotation import _ANNOTATION_TYPES
 from kolena.workflow.asset import _ASSET_TYPES
