+++ tests/unit/workflow/test_evaluator.py
@@ -23,7 +23,7 @@ import pydantic
 import pytest
 
 from kolena._experimental.workflow.thresholded import ThresholdedMetrics
-from kolena.workflow._datatypes import DataObject
+from kolena._utils.datatypes import DataObject
 from kolena.workflow.annotation import BoundingBox
 from kolena.workflow.annotation import Polyline
 from kolena.workflow.evaluator import MetricsTestCase
