+++ tests/unit/workflow/test_ground_truth.py
@@ -22,9 +22,9 @@ import pytest
 from .data import ComplexBoundingBox
 from .data import ImageTriplet
 from .data import NestedComplexBoundingBox
+from kolena._utils.datatypes import DataObject
 from kolena.workflow import GroundTruth
 from kolena.workflow import TestSample
-from kolena.workflow._datatypes import DataObject
 from kolena.workflow.annotation import BoundingBox
 from kolena.workflow.annotation import Keypoints
 from kolena.workflow.annotation import LabeledBoundingBox
